Work on ipc; better logging
Apparently the "fern" logging crate isn't so popular now, but I spent forever trying to get "tracing" to do the equivalent tasks and couldn't figure it out.

https://github.com/daboross/fern/discussions/147

Async tasks/threading are STILL being a nightmare. How is it so much harder to do this than it is to just control jobs with & and wait in the shell?